requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via starlette
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via click
deprecated==1.2.14
    # via limits
fastapi==0.112.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.7
    # via anyio
importlib-resources==6.4.0
    # via limits
limits==3.13.0
    # via slowapi
packaging==24.1
    # via limits
pillow==10.4.0
    # via -r requirements.in
pydantic==2.8.2
    # via fastapi
pydantic-core==2.20.1
    # via pydantic
slowapi==0.1.9
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.37.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   limits
    #   pydantic
    #   pydantic-core
uvicorn==0.30.5
    # via -r requirements.in
wrapt==1.16.0
    # via deprecated
